NURBS gravity column in 3D
I  had some trouble initially to get the code to understand that one particle left the element and went into another. Part of the problem is that we use sides to EDGES what new elements are adjacent to the old element. This made it quite difficult when we were trying to use the sub-tetrahedral identifier to find what side we are crossing. I managed to break down the hexahedron into 4 tetrahedrons. I also had an extra check to see if we are inside the element when we update the new element. It was gibing not true for the .IsInside. logical variable. Anyway, it could be improved  but it is alright for now. Thanks for reading :)